pyproject.toml
[project]
name = "agenticlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.10",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.6",
    "langsmith>=0.4.7",
    "python-dotenv>=1.1.1",
]
